Move tests in `router.test.ts` to more scoped test files
